+++ docs/history.md
@@ -1,5 +1,10 @@
 # Racing+ Version History and News
 
+### *v0.74.19* - January 2nd, 2022
+
+- Fixed the bug where in seeded races, Card Reading portals would not be removed properly.
+- Fixed the bug where the buttons would not spawn under certain circumstances. (Thanks cutifire)
+
 ### *v0.74.18* - December ?th, 2021
 
 - Vanishing Twin will now work faster. (It is reimplemented from scratch.)